moving precision to signed, since the difference could be negative

+++ dango/dex/src/cron.rs
@@ -893,7 +893,7 @@ fn clear_orders_of_pair(
 
             // The spread absolute needs to be adjusted according to difference in the tokens's precision.
             spread_absolute_f64 *=
-                10.0_f64.powi((base_price.precision() - quote_price.precision()) as i32);
+                10.0_f64.powi(base_price.precision() as i32 - quote_price.precision() as i32);
 
             let spread_percentage_f64: f64 =
                 spread_absolute.checked_div(mid)?.to_string().parse()?;
